Split ApiMismatch tests into GET and PUT
