+++ tests/Feature/ApiMismatchRouteGetTest.php
@@ -9,9 +9,8 @@ use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\WithFaker;
-use Laravel\Sanctum\Sanctum;
 
-class ApiMismatchRouteTest extends TestCase
+class ApiMismatchRouteGetTest extends TestCase
 {
     use RefreshDatabase, WithFaker;
 
@@ -208,77 +207,6 @@ class ApiMismatchRouteTest extends TestCase
             );
     }
 
-    public function test_put_review_status_returns_updated_mismatch()
-    {
-        $mismatch = $this->generateSingleMismatch();
-
-        $reviewer = User::factory()->create();
-        $review_status = 'wikidata';
-
-        Sanctum::actingAs($reviewer);
-        $this->travelTo(now()); // freezes time to ensure correct updated_at timestamp
-
-        $mismatchId =  $mismatch->id;
-        $response = $this->json(
-            'PUT',
-            self::MISMATCH_ROUTE . '/' . $mismatchId,
-            [ 'review_status' => $review_status ]
-        );
-
-        $response->assertSuccessful()
-            ->assertJsonStructure(
-                [  // array of mismatch items
-                    'id',
-                    'item_id',
-                    'statement_guid',
-                    'property_id',
-                    'wikidata_value',
-                    'external_value',
-                    'external_url',
-                    'review_status',
-                    'reviewer' => [],
-                    'import' => [
-                        'id',
-                        'status',
-                        'description',
-                        'created',
-                        'expires',
-                        'uploader' => [
-                            'id',
-                            'username',
-                            'mw_userid'
-                        ]
-                    ]
-                ]
-            )->assertJson(
-                [
-                    'review_status' => $review_status,
-                    'reviewer' => [
-                        'id' => $reviewer->id,
-                        'username' => $reviewer->username,
-                        'mw_userid' => $reviewer->mw_userid
-                    ],
-                    'updated_at' => now()->millisecond(0)->toISOString()
-                ]
-            );
-    }
-
-    public function test_put_review_status_with_invalid_mismatch_id_returns_validation_error()
-    {
-
-        Sanctum::actingAs(User::factory()->create());
-        $review_status = 'wikidata';
-
-        $mismatchId =  'not_a_valid_id';
-        $response = $this->json(
-            'PUT',
-            self::MISMATCH_ROUTE . '/' . $mismatchId,
-            [ 'review_status' => $review_status ]
-        );
-
-        $response->assertStatus(404);
-    }
-
     public function test_missing_item_ids_returns_validation_error()
     {
         $response = $this->json('GET', self::MISMATCH_ROUTE);  // ids missing
@@ -320,98 +248,6 @@ class ApiMismatchRouteTest extends TestCase
                 'attribute' => 'ids.0'
             ])
         ]);
-    }
-
-    public function test_unauthenticated_put_returns_401()
-    {
-        $response = $this->json(
-            'PUT',
-            self::MISMATCH_ROUTE . '/1',
-            [ 'review_status' => 'wikidata' ]
-        );
-
-        $response->assertStatus(401);
-    }
-
-    public function test_put_non_existing_mismatch_returns_404()
-    {
-        Sanctum::actingAs(User::factory()->create());
-        $response = $this->json(
-            'PUT',
-            self::MISMATCH_ROUTE . '/31415',
-            [ 'review_status' => 'wikidata' ]
-        );
-
-        $response->assertStatus(404);
-    }
-
-    public function test_invalid_review_status_returns_validation_error()
-    {
-        Sanctum::actingAs(User::factory()->create());
-        $mismatch =  $this->generateSingleMismatch();
-        $response = $this->json(
-            'PUT',
-            self::MISMATCH_ROUTE . '/' . $mismatch->id,
-            [ 'review_status' => 'potato' ]
-        );
-
-        $response->assertJsonValidationErrors([
-                'review_status' => __('validation.in', [
-                    'attribute' => 'review status'
-                ])
-            ]);
-    }
-
-    public function fillablePropertyOtherThanReviewStatus()
-    {
-        return [
-            ['property_id' , 'P1234'],
-            ['statement_guid', 'Q111$1234'],
-            ['property_id', 'P666'],
-            ['wikidata_value', 'Potato'],
-            ['external_value', 'Tomato'],
-            ['external_url', 'http://potato.com']
-        ];
-    }
-
-    /**
-     *  Test parameters other than review status
-     *
-     *  @return void
-     *  @dataProvider fillablePropertyOtherThanReviewStatus
-     */
-    public function test_parameter_other_than_review_status_in_put_request_returns_validation_error($key, $parameter)
-    {
-        Sanctum::actingAs(User::factory()->create());
-        $mismatch = $this->generateSingleMismatch();
-        $response = $this->json(
-            'PUT',
-            self::MISMATCH_ROUTE . '/' . $mismatch->id,
-            [
-                'review_status' => 'wikidata',
-                 $key => $parameter
-            ]
-        );
-
-        $response->assertJsonValidationErrors([
-                $key => __('validation.prohibited', [
-                    'attribute' => str_replace('_', ' ', $key)
-                ])
-            ]);
-    }
-
-    private function generateSingleMismatch()
-    {
-
-        $import = ImportMeta::factory()
-        ->for(User::factory()->uploader())
-        ->create();
-
-        $mismatch = Mismatch::factory()
-            ->for($import)
-            ->create();
-
-        return $mismatch;
     }
 
     private function seedMismatches()
